% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autocorrelated_genes.R
\name{autocorrelated_genes}
\alias{autocorrelated_genes}
\title{autocorrelated_genes()}
\usage{
autocorrelated_genes(
  obj,
  alternative = "greater",
  adjustPv = TRUE,
  alpha = 0.05,
  minPercentCells = 0.05,
  verbose = TRUE
)
}
\arguments{
\item{obj}{STew S3 Object}

\item{alternative}{a character string specifying the alternative hypothesis, chosen from one of "greater", "two-sided" or "less". Default is "greater"}

\item{adjustPv}{Adjusted P value}

\item{alpha}{P-value threshold for LISA score to be considered significant}

\item{minPercentCells}{Minimum percent of cells that must be driving spatial pattern}

\item{verbose}{A logical parameter that controls the amount of output or information displayed during the execution of the function. When verbose = TRUE, the function will provide more detailed and informative output during its execution and when verbose = FALSE (default setting), the function will produce minimal output.}
}
\value{
Subset regularized gene expression matrix by filtered genes
}
\description{
Filtering gene expression matrix by spatially auto-correlated genes and correction for multiple-testing
}
