% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildSNNSeurat.R
\name{BuildSNNSeurat}
\alias{BuildSNNSeurat}
\title{BuildSNNSeurat()}
\usage{
BuildSNNSeurat(obj, k.param = 30, prune.SNN = 1/15, nn.eps = 0, cols = NULL)
}
\arguments{
\item{obj}{stCCA S3 Object}

\item{k.param}{Maximum number of nearest neighbors to compute. Default - 30.}

\item{prune.SNN}{Sets the cutoff for acceptable Jaccard index when computing the neighborhood overlap for the SNN construction. Any edges with values less than or equal to this will be set to 0 and removed from the SNN graph. Essentially sets the stringency of pruning (0 --- no pruning, 1 --- prune everything). Default - 1/15.}

\item{nn.eps}{Error bound: default of 0.0 implies exact nearest neighbor search}

\item{cols}{columns selected}
}
\description{
Build SNN matrix by inputting canonical variates of spatial-driven canonical loading
}
