% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cca_main.R
\name{cca_main}
\alias{cca_main}
\title{cca_main()}
\usage{
cca_main(
  x,
  z,
  obj = NULL,
  penaltyx = NULL,
  penaltyz = NULL,
  K = 1,
  niter = 15,
  v = NULL,
  trace = TRUE,
  standardize = TRUE,
  xnames = colnames(x),
  znames = colnames(z),
  upos = FALSE,
  uneg = FALSE,
  vpos = FALSE,
  vneg = FALSE,
  outcome = NULL,
  y = NULL,
  cens = NULL
)
}
\arguments{
\item{x}{Expression-driven adjacency matrix}

\item{z}{Spatial-driven adjacency matrix}

\item{obj}{stCCA S3 Object}

\item{penaltyx}{The penalty to be applied to the matrix x, i.e. the penalty that results in the canonical vector u. If typex is "standard" then the L1 bound on u is penaltyx*sqrt(ncol(x)). In this case penaltyx must be between 0 and 1 (larger L1 bound corresponds to less penalization).}

\item{penaltyz}{The penalty to be applied to the matrix z, i.e. the penalty that results in the canonical vector v. If typez is "standard" then the L1 bound on v is penaltyz*sqrt(ncol(z)). In this case penaltyz must be between 0 and 1 (larger L1 bound corresponds to less penalization).}

\item{K}{The number of u's and v's desired; that is, the number of canonical vectors to be obtained.}

\item{niter}{The number of iterations to be performed. Default - 20.}

\item{v}{-desc}

\item{trace}{-desc}

\item{standardize}{-desc}

\item{xnames}{-desc}

\item{znames}{-desc}

\item{upos}{-desc}

\item{uneg}{-desc}

\item{vpos}{-desc}

\item{vneg}{-desc}

\item{outcome}{-desc}

\item{y}{-desc}

\item{cens}{-desc}
}
\description{
Perform Sparse CCA using the optimal tuning parameters
}
