% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_avg_plots.R
\name{create_avg_plots}
\alias{create_avg_plots}
\title{create_avg_plots()}
\usage{
create_avg_plots(mean_var, obs_mean_var, models, log_transform = FALSE)
}
\arguments{
\item{mean_var}{A data frame containing the mean and variance data.}

\item{obs_mean_var}{A data frame containing the observed mean and variance data.}

\item{models}{A character vector specifying the models.}

\item{log_transform}{A logical value indicating whether to perform log transformation on the data.}
}
\value{
A list of average plots.
}
\description{
Create Average Plots
}
\details{
This function creates average plots based on the mean and variance data.
}
