% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_adjacency_matrix.R
\name{expression_adjacency_matrix}
\alias{expression_adjacency_matrix}
\title{To calculate cell-to-cell similarity matrix on gene (spatially auto-coorelated genes) expression based on Euclidean distance; to build up a K-nearest neighbors (KNN) graph based on the cell-to-cell similarity matrix; then to convert the KNN neighbor relation matrix into an adjacency matrix}
\usage{
expression_adjacency_matrix(
  obj,
  nk = 200,
  symm = FALSE,
  weight = TRUE,
  mode = "undirected",
  weighted = FALSE
)
}
\arguments{
\item{obj}{stCCA S3 Object}

\item{nk}{Number of neighbors}

\item{symm}{If TRUE, the connectivity matrix is made symmetrical. Default - FALSE}

\item{weight}{If TRUE, the weighted KNN graph is created. Default - FALSE.}

\item{mode}{A character string chosen from c(""directed", "undirected", "max", "min", "upper", "lower", "plus")}

\item{weighted}{To specify whether to create a weighted graph from an adjacency matrix. If NULL, an unweighted graph is created. If TRUE, a weighted graph is created.}
}
\description{
To calculate cell-to-cell similarity matrix on gene (spatially auto-coorelated genes) expression based on Euclidean distance; to build up a K-nearest neighbors (KNN) graph based on the cell-to-cell similarity matrix; then to convert the KNN neighbor relation matrix into an adjacency matrix
}
