% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindClusters.R
\name{FindClusters}
\alias{FindClusters}
\title{FindClusters()}
\usage{
FindClusters(resolution_list, obj)
}
\arguments{
\item{resolution_list}{List of resolutions to use for clustering}

\item{obj}{stCCA S3 Object}
}
\description{
Perform cell clustering using different resolutions
}
