% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_stats.R
\name{get_model_stats}
\alias{get_model_stats}
\title{get_model_stats()}
\usage{
get_model_stats(effect_plot, dt, models)
}
\arguments{
\item{effect_plot}{The effect plot containing the predicted values from different models.}

\item{dt}{The data frame containing the observed values.}

\item{models}{A character vector specifying the models to calculate statistics for.}
}
\value{
A list with two elements:
\itemize{
\item mean_var: A data frame with the mean and variance values for each gene and model.
\item obs_mean_var: A data frame with the observed mean and variance values for each gene.
}
}
\description{
Get Model Statistics
}
\details{
This function calculates the model statistics for a given effect plot and data frame.
}
