% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint_embedding.R
\name{joint_embedding}
\alias{joint_embedding}
\title{joint_embedding()}
\usage{
joint_embedding(obj)
}
\arguments{
\item{obj}{stCCA S3 Object}
}
\description{
Extracting spatial-informed canonical loading
}
