% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multimodel_count_regression.R
\name{multimodel_count_regression}
\alias{multimodel_count_regression}
\title{multimodel_count_regression()}
\usage{
multimodel_count_regression(
  outcomes_interest,
  models,
  plot_density,
  plot_qq,
  estimate,
  avg_effect,
  rootogram,
  full_model,
  null_model,
  data,
  workers,
  path
)
}
\arguments{
\item{outcomes_interest}{A character vector specifying the outcomes of interest.}

\item{models}{A character vector specifying the models to be fitted.}

\item{plot_density}{A logical value indicating whether to plot density plots of the fitted models.}

\item{plot_qq}{A logical value indicating whether to plot Q-Q plots of the fitted models.}

\item{estimate}{A logical value indicating whether to estimate model parameters and standard errors.}

\item{avg_effect}{A logical value indicating whether to compute average marginal effects.}

\item{rootogram}{A logical value indicating whether to generate rootograms for the fitted models.}

\item{full_model}{The formula for the full model.}

\item{null_model}{The formula for the null model.}

\item{data}{The data frame containing the variables of interest.}

\item{workers}{The number of parallel workers to use for parallel computing.}

\item{path}{The path to save the plots.}
}
\value{
A list containing the results of the multimodel count regression analysis.
}
\description{
Multimodel Count Regression
}
\details{
This function performs multimodel count regression analysis.
}
