% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_cca_permute.R
\name{parallel_cca_permute}
\alias{parallel_cca_permute}
\title{Selecting tuning parameters for sparse CCA}
\usage{
parallel_cca_permute(
  x,
  z,
  obj,
  penaltyxs = NULL,
  penaltyzs = NULL,
  niter = 3,
  v = NULL,
  trace = FALSE,
  nperms = 25,
  standardize = TRUE,
  upos = FALSE,
  uneg = FALSE,
  vpos = FALSE,
  vneg = FALSE,
  outcome = NULL,
  y = NULL,
  cens = NULL
)
}
\arguments{
\item{x}{Expression-driven adjacency matrix}

\item{z}{Spatial-driven adjacency matrix}

\item{obj}{stCCA S3 Object}

\item{penaltyxs}{-desc}

\item{penaltyzs}{-desc}

\item{niter}{The number of iterations to be performed. Default - 20.}

\item{v}{-desc}

\item{trace}{-desc}

\item{nperms}{-desc}

\item{standardize}{-desc}

\item{upos}{-desc}

\item{uneg}{-desc}

\item{vpos}{-desc}

\item{vneg}{-desc}

\item{outcome}{-desc}

\item{y}{-desc}

\item{cens}{-desc}
}
\description{
Selecting tuning parameters for sparse CCA
}
\references{
We adapted and modified the code from the CCA function in the PMA package to be compatible and scalable with single-cell multimodal integration through parallel computing.
}
