% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cluster.R
\name{plot_cluster}
\alias{plot_cluster}
\title{plot_cluster()}
\usage{
plot_cluster(coordis, label, colors, t = NULL, subt = NULL)
}
\arguments{
\item{coordis}{Spatial location of cells with cell barcode as rows and coordinates as columns.}

\item{label}{Cluster label as a column of data frame.}

\item{colors}{Color palette for clusters.}

\item{t}{Plot Title. Null in default.}

\item{subt}{Plot subtitle. Null in default.}
}
\description{
Visualizing clusters in human brain image
}
