% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_corr.R
\name{plot_corr}
\alias{plot_corr}
\title{Barplot of canonical correlation for each pair of canonical variate}
\usage{
plot_corr(cc, t, tsub = NULL)
}
\arguments{
\item{cc}{Output of CCA() function. Result of Sparse CCA.}

\item{t}{Title}

\item{tsub}{Subtitle}
}
\description{
Barplot of canonical correlation for each pair of canonical variate
}
