% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sct.R
\name{sct}
\alias{sct}
\title{sct()}
\usage{
sct(obj, n_genes = 2000)
}
\arguments{
\item{obj}{stCCA S3 Object}

\item{n_genes}{Number of genes}
}
\description{
Apply variance stabilizing transformation to UMI count data using a regularized Negative Binomial regression model. This will remove unwanted effects from UMI data
}
