% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_adjacency_matrix.R
\name{spatial_adjacency_matrix}
\alias{spatial_adjacency_matrix}
\title{spatial_adjacency_matrix()}
\usage{
spatial_adjacency_matrix(obj)
}
\arguments{
\item{obj}{stCCA S3 Object}
}
\value{
A matrix that describes the adjacency relationships between cells. The entries in the matrix are binary values (0 or 1) that indicate whether cells are adjacent to each other or not.
}
\description{
Creating spatial-driven adjacency matrix using cell spatial coordinates
}
