% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_gradient.R
\name{spatial_gradient}
\alias{spatial_gradient}
\title{spatial_gradient()}
\usage{
spatial_gradient(obj, t = NULL)
}
\arguments{
\item{obj}{stCCA S3 Object}

\item{t}{A character string to define plot title. Null in default.}
}
\description{
Spatial-informed cell gradient plots
}
