% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stew_Obj.R
\name{STew_Obj}
\alias{STew_Obj}
\title{STew S3 object}
\usage{
STew_Obj(count_exp = NULL, spatial = NULL)
}
\arguments{
\item{count_exp}{count_exp data}

\item{spatial}{spatial data}
}
\value{
S3 Object
}
\description{
STew S3 object
}
