% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_genes.R
\name{var_genes}
\alias{var_genes}
\title{Calculating row mean for each gene using normalized gene expression matrix}
\usage{
var_genes(obj)
}
\arguments{
\item{obj}{stCCA object}
}
\description{
Calculating row mean for each gene using normalized gene expression matrix
}
